Improve query join calls
When making certain queries, we want to ensure we are making nested join calls
to limit our subset in order to return the correct result.

Correctly filter data by using a call to `DISTINCT` within a call to `having`
to ensure duplicate rows are not used in the calculation.

Issues
------
Closes #38